# backend-runtime/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar actividades y sus checks diarios.",
        "title": "API de Actividades",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/actividades": {
            "get": {
                "description": "Devuelve las actividades activas, ordenadas por id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actividades"
                ],
                "summary": "Listar actividades",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Actividad"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actividades"
                ],
                "summary": "Crear actividad",
                "parameters": [
                    {
                        "description": "Datos de la actividad",
                        "name": "actividad",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CrearActividadInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Actividad"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/actividades/{id}": {
            "put": {
                "description": "Reemplaza título, descripción, categoría y prioridad de una actividad existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actividades"
                ],
                "summary": "Editar actividad",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la actividad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la actividad",
                        "name": "actividad",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ActualizarActividadInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Actividad"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Borrado lógico: marca la actividad como inactiva (activa = FALSE)\nen vez de borrar la fila, para no perder su historial de checks_diarios\n(que referencia actividad_id) ni su aporte pasado al resumen.",
                "tags": [
                    "actividades"
                ],
                "summary": "Eliminar actividad",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la actividad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categorias": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Listar categorías",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Categoria"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Crear categoría",
                "parameters": [
                    {
                        "description": "Nombre de la categoría",
                        "name": "categoria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CrearCategoriaInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Categoria"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categorias/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Renombrar categoría",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la categoría",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo nombre",
                        "name": "categoria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ActualizarCategoriaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Categoria"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "categorias"
                ],
                "summary": "Eliminar categoría",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la categoría",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/checks": {
            "post": {
                "description": "Solo se permite un check por actividad y por día (fecha del\nservidor); si ya existe uno, devuelve 409.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checks"
                ],
                "summary": "Dar check a una actividad hoy",
                "parameters": [
                    {
                        "description": "Datos del check",
                        "name": "check",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CrearCheckInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.CheckDiario"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Ya existe un check para esta actividad hoy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resumen": {
            "get": {
                "description": "Usa la vista resumen_actividades",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resumen"
                ],
                "summary": "Resumen de cumplimiento por actividad",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ResumenActividad"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Actividad": {
            "type": "object",
            "properties": {
                "activa": {
                    "type": "boolean"
                },
                "actualizado_en": {
                    "type": "string"
                },
                "categoria_id": {
                    "type": "integer"
                },
                "categoria_nombre": {
                    "type": "string"
                },
                "check_hoy": {
                    "description": "CheckHoy indica si esta actividad ya tiene un check_diario registrado en\nla fecha actual del servidor (checks_diarios tiene UNIQUE(actividad_id, fecha),\nasí que como máximo puede haber uno). Se lo mandamos ya resuelto al frontend\npara que no tenga que inferir el estado del día a partir de clicks locales.",
                    "type": "boolean"
                },
                "creado_en": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "prioridad": {
                    "type": "integer"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "main.ActualizarActividadInput": {
            "type": "object",
            "properties": {
                "categoria_id": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "prioridad": {
                    "type": "integer"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "main.ActualizarCategoriaInput": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                }
            }
        },
        "main.Categoria": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "main.CrearActividadInput": {
            "type": "object",
            "properties": {
                "categoria_id": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "prioridad": {
                    "type": "integer"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "main.CrearCategoriaInput": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                }
            }
        },
...
    }
}